app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
neo4j==5.16.0
pandas==2.2.0
